trying to get pickle of our datatypes to work
pickling would make it easier to hack on the event_search by reading
pre-searched hits from disk rather than running the full pipeline
every time

It mostly works for the simpler types but the following is causing some
issue:
* enum (flagger_values)
* unclear how to handle the inheritance of scan <-> filterbank which may
  not strictly be necessary for early hacking

+++ bliss/core/include/core/hit.hpp
@@ -14,8 +14,8 @@ struct hit {
     int64_t rate_index;
     float   drift_rate_Hz_per_sec;
     float   snr;
-    int64_t bandwidth;
-    double  binwidth;
+    double  bandwidth;
+    int64_t binwidth;
     rfi     rfi_counts;
 };
